added resources directory with functions and images files, fixed off by one error
