+++ resources/ascii_images.mjs
@@ -1,0 +1,41 @@
+//function to display title
+export function title () {
+    return console.log(`
+        ____  __ __    ___  _____ _____  
+       /    ||  |  |  /  _]/ ___// ___/  
+      |   __||  |  | /  [_(   \\_(   \\_   
+      |  |  ||  |  ||    _]\\__  |\\__  |  
+      |  |_ ||  :  ||   [_ /  \\ |/  \\ |  
+      |     ||     ||     |\\    |\\    |  
+      |___,_| \\__,_||_____| \\___| \\___|  
+                                         
+      __    __  __ __   ____  ______  _____ 
+     |  |__|  ||  |  | /    ||      |/     |
+     |  |  |  ||  |  ||  o  ||      ||  Y  |
+     |  |  |  ||  _  ||     ||_|  |_||__|  |
+     |  '  '  ||  |  ||  _  |  |  |     |__|
+      \\      / |  |  ||  |  |  |  |      __ 
+       \\_/\\_/  |__|__||__|__|  |__|     |__|
+                                    `);
+};
+
+//function to display winning image
+export function win () {
+    return console.log(`\n
+                 __ __   ___   __ __       
+                |  |  | /   \\ |  |  |      
+                |  |  ||     ||  |  |      
+                |  ~  ||  O  ||  |  |      
+                |___, ||     ||  :  |      
+                |     ||     ||     |      
+                |____/  \\___/  \\__,_|      
+                                           
+             __    __  ____  ____       __ 
+            |  |__|  ||    ||    \\     |  |
+            |  |  |  | |  | |  _  |    |  |
+            |  |  |  | |  | |  |  |    |__|
+            |  '  '  | |  | |  |  |     __ 
+             \\      /  |  | |  |  |    |  |
+              \\_/\\_/  |____||__|__|    |__|
+                                           `);
+};
